# _backend/openapi.yaml
openapi: 3.0.0
info:
  title: CatTok
  description: >
    Made for **CatTok** frontend client, GitHub [littlefuny/cattok](https://github.com/littlefunky/cattok)

    <iframe src="https://giphy.com/embed/JuFwy0zPzd6jC" width="480" height="380" frameBorder="0" class="giphy-embed" allowFullScreen></iframe><p><a href="https://giphy.com/gifs/computer-computers-caturday-JuFwy0zPzd6jC">via GIPHY</a></p>
      
    This project was made for educational purpose only 🚀.
  version: 1.0.0
servers:
  - url: /v1
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: jwt
...
  responses:
    UnexpectedError:
      description: Unexpected error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedError:
      description: Unauthorized error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Fail"
          example:
            status: fail
            data: "Unauthorized"
  schemas:
...
    Success:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          default: success
        data:
          description: Returned data
          nullable: true
    Fail:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          default: fail
        data:
          description: Failed message
    Error:
      type: object
      required:
        - status
        - data
      properties:
        status:
          type: string
          default: error
        message:
          description: Error message
          type: string
...
    Post:
      type: object
      required:
        - id
        - user_id
        - title
        - description
        - video_url
        - public_date
        - heart
      properties:
        id:
          type: string
        user_id:
          type: string
        title:
          type: string
        description:
          type: string
        video_url:
          type: string
        public_date:
          type: string
          format: date-time
        heart:
          type: integer
    Comment:
      type: object
      required:
        - id
        - user_id
        - comment
        - comment_date
      properties:
        id:
          type: string
        user_id:
          type: string
        comment:
          type: string
        comment_date:
          type: string
          format: date-time
...
paths:
...
  /post:
    get:
      tags:
        - Post
      description: Get list of `Post`
      security: []
      parameters:
        - name: cursor
          required: false
          in: query
          schema:
            type: string
        - name: limit
          required: false
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Post"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
    post:
      tags:
        - Post
      description: Create new post
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - description
                - video
              properties:
                title:
                  type: string
                description:
                  type: string
                video:
                  type: string
                  format: binary
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          post_id:
                            type: string
        "400":
          description: Malformed or missing required data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /post/{post_id}:
    get:
      tags:
        - Post
      description: Get single `Post` from ID
      security: []
      parameters:
        - name: post_id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Post"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
    put:
      tags:
        - Post
      description: Update post
      parameters:
        - name: post_id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                video:
                  type: string
                  format: binary
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Malformed or missing required data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: No permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
    delete:
      tags:
        - Post
      description: Delete post
      parameters:
        - name: post_id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: No permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /post/{post_id}/heart:
    post:
      tags:
        - Post
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
      description: Give heart to a post
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /post/{post_id}/comment:
    get:
      tags:
        - Post
        - Comment
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
      description: Get list of `Comment` of post
      security: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Comment"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
    post:
      tags:
        - Post
        - Comment
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
              properties:
                comment:
                  type: string
      description: Create new comment
      responses:
        "201":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          comment_id:
                            type: string
        "400":
          description: Malformed or missing required data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Post not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
  /post/{post_id}/comment/{comment_id}:
    get:
      tags:
        - Post
        - Comment
      security: []
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
      description: Get specific `Comment` by ID
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Success"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Comment"
        "404":
          description: Post or comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
    put:
      description: Update comment
      tags:
        - Post
        - Comment
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - comment
              properties:
                comment:
                  type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "400":
          description: Malformed or missing required data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: No permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "404":
          description: Post or comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"
    delete:
      tags:
        - Post
        - Comment
      parameters:
        - name: post_id
          in: path
          required: true
          schema:
            type: string
        - name: comment_id
          in: path
          required: true
          schema:
            type: string
      description: Delete comment
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Success"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: No permission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        "404":
          description: Post or comment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fail"
        default:
          $ref: "#/components/responses/UnexpectedError"